+++ cmake/FindDAGMC.cmake
@@ -12,7 +12,7 @@ message(STATUS ${DAGMC_ROOT})
 find_path(DAGMC_CMAKE_CONFIG NAMES DAGMCConfig.cmake
           HINTS ${DAGMC_ROOT} $ENV{DAGMC_ROOT}
           PATHS ENV LD_LIBRARY_PATH
-          PATH_SUFFIXES lib Lib cmake lib/cmake/
+          PATH_SUFFIXES lib Lib cmake lib/cmake/dagmc
           NO_DEFAULT_PATH)
 
 message(STATUS "Found DAGMC in ${DAGMC_CMAKE_CONFIG}")